+++ support-frontend/assets/helpers/deliveryCheck.js
